refactor: extract authentication logic into custom hooks
- Create useLogin hook for login page functionality
- Create useRegister hook for registration page functionality
- Move form state management to respective hooks
- Move API calls and error handling to hooks
- Clean up component files by removing direct state management

This change improves code organization by separating concerns
and making authentication logic more reusable across the application.
